Los componentes Productos, Cliente y Proveedor ya estan validados y con mensajes de confirmacion. Avisar si se quiere modificar algo o si algo esta mal plox.

+++ backend/frontend/src/components/cliente/ClienteDelete.js
@@ -5,19 +5,30 @@ export default class ClienteDelete extends Component {
         super(props);
         this.state = {
             data: [],
+            dataTable: [],
             show: false,
             clientId: 0,
             nombre_C: "",
-            balance: 0.0
+            balance: 0.0,
+            buscador: ""
         };
         this.getClientData = this.getClientData.bind(this);
         this.deleteData = this.deleteData.bind(this);
-        this.modiData = this.modiData.bind(this);
-        this.applyChanges = this.applyChanges.bind(this);
+    
 
-        this.getclientId = this.getclientId.bind(this)
-        this.getnombre_C = this.getnombre_C.bind(this)
-        this.getbalance = this.getbalance.bind(this)
+        this.getclientId = this.getclientId.bind(this);
+        this.getnombre_C = this.getnombre_C.bind(this);
+        this.getbalance = this.getbalance.bind(this);
+
+        this.buscar = this.buscar.bind(this);
+        this.getBuscador = this.getBuscador.bind(this);
+    }
+
+    getBuscador(e) {
+        this.setState({
+            buscador: e.target.value,
+        });
+        this.buscar(e);
     }
 
     getclientId(value) {
@@ -38,12 +49,14 @@ export default class ClienteDelete extends Component {
        });
     }
 
+
     getClientData() {
         fetch('/cliente/get-client')
             .then(response => response.json())
             .then((data) => {
                 this.setState({
-                    data: data
+                    data: data,
+                    dataTable: data,
                 });
             });
     }
@@ -52,53 +65,43 @@ export default class ClienteDelete extends Component {
         this.getClientData();
     }
 
-    deleteData(clientId) {
-        fetch('/cliente/delete-cliente/' + clientId, {
-            method: 'DELETE',
-            body: JSON.stringify(this.state),
-        }).then(response => response)
+    deleteData(clientId, nombre_C) {
+        if (confirm("Desea eliminar al cliente: " + "  " + nombre_C)) {
+            fetch('/cliente/delete-cliente/' + clientId, {
+                method: 'DELETE',
+                body: JSON.stringify(this.state),
+            }).then(response => response)
             .then((data) => {
                 if (data) {
                     this.getClientData();
                 }
             });
+            alert("Cliente eliminado con exito");
+            this.setState({buscador: ""});
+        }
     }
 
-    modiData(clientId) {
-        fetch('/cliente/update-cliente/' + clientId)
-        .then(response => response.json())
-        .then((data) => {
-            //console.log(clientId)
-            this.setState({
-                clientId: clientId,
-                nombre_C: data.nombre_C,
-                balance: data.balance
-            });
+    buscar(e) {
+        const nombre = e.target.value.toLowerCase();
+        const auxData = []
+        for (let i = 0; i < this.state.data.length; i++) {
+            const element = this.state.data[i];
+            const str = element.nombre_C.toLowerCase();
+            if (str.includes(nombre)) {
+                auxData.push(element);    
+            }
+        }
+        this.setState({
+            dataTable: auxData,
         });
-        { this.setState({ show: true }) }
-    }
-
-    applyChanges(clientId) {
-        const requiestClient = {
-            method: 'PUT',
-            headers: { "Content-Type": "application/json" },
-            body: JSON.stringify({
-                clientId: this.state.clientId,
-                nombre_C: this.state.nombre_C,
-                balance: this.state.balance
-            }),
-        };
-        fetch("/cliente/update-cliente/"+ clientId, requiestClient)
-        .then((response) => response.json());
-        { this.setState({ show: false }) }
     }
 
     render() {
-        const clienData = this.state.data;
+        const clienData = this.state.dataTable;
         const rows = clienData.map((clien) =>
             <tr key={clien.clientId}>
-                <td onClick={() => this.deleteData(clien.clientId)} >{clien.nombre_C}</td>
-                <td onClick={() => this.deleteData(clien.clientId)} >{clien.balance}</td>
+                <td onClick={() => this.deleteData(clien.clientId, clien.nombre_C)} >{clien.nombre_C}</td>
+                <td onClick={() => this.deleteData(clien.clientId, clien.nombre_C)} >{clien.balance}</td>
             </tr>
         );
 
@@ -108,7 +111,7 @@ export default class ClienteDelete extends Component {
                 </h2>
                 <form>
                     <div className="group">
-                        <input type="text" required />
+                        <input type="text" required value={this.state.buscador} onChange={e => this.getBuscador(e)} />
                         <span className="highlight"></span>
                         <span className="bar"></span>
                         <label>Nombre</label>
